fix: 출석 수 세는 부분 long -> int 타입 변경

+++ dolharubang/src/main/java/com/dolharubang/repository/AttendanceRepository.java
@@ -9,6 +9,6 @@ public interface AttendanceRepository extends JpaRepository<Attendance, Long> {
 
     boolean existsByMemberAndDate(Member member, LocalDate date);
 
-    long countByMember(Member member);
+    int countByMember(Member member);
 
 }
